Finished refactor, now the processor only puts the items to their correlative repository. Introduced one final background-running interactor: ProcessMessageInteractor.
