+++ app/src/main/java/com/teamagam/gimelgimel/app/injectors/components/ApplicationComponent.java
@@ -36,6 +36,7 @@ import com.teamagam.gimelgimel.domain.map.repository.DisplayedEntitiesRepository
 import com.teamagam.gimelgimel.domain.map.repository.GeoEntitiesRepository;
 import com.teamagam.gimelgimel.domain.map.repository.SelectedEntityRepository;
 import com.teamagam.gimelgimel.domain.map.repository.SingleDisplayedItemRepository;
+import com.teamagam.gimelgimel.domain.messages.ProcessMessagesInteractor;
 import com.teamagam.gimelgimel.domain.messages.UpdateUnreadCountInteractor;
 import com.teamagam.gimelgimel.domain.messages.poller.StartFetchingMessagesInteractor;
 import com.teamagam.gimelgimel.domain.messages.repository.MessagesRepository;
@@ -110,6 +111,8 @@ public interface ApplicationComponent {
   VectorLayersVisibilityRepository vectorLayersVisibilityRepository();
 
   StartFetchingMessagesInteractor startFetchingMessagesInteractor();
+
+  ProcessMessagesInteractor processMessagesInteractor();
 
   UpdateUnreadCountInteractor updateUnreadCountInteractor();
